pyproject.toml
[tool.poetry]
name = "django5course"
version = "0.1.0"
description = ""
authors = ["Sean House <33880135+sean-house@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
django = "^5.0.1"
django-allauth = {extras = ["socialaccount"], version = "^65.2.0"}
psycopg = "^3.2.3"
dj-database-url = "^2.3.0"
django-anymail = {extras = ["mailgun"], version = "^12.0"}
django-widget-tweaks = "^1.5.0"
gunicorn = "^23.0.0"
whitenoise = {extras = ["brotli"], version = "^6.8.2"}
markdown = "^3.7"


[tool.poetry.group.dev.dependencies]
django-debug-toolbar = "^4.4.6"
django-browser-reload = "^1.17.0"
pytest = "^8.3.4"
pytest-django = "^4.9.0"
playwright = "^1.49.1"
pytest-playwright = "^0.6.2"
pytest-xdist = "^3.6.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff.lint]
select = ["ALL"]

[tool.ruff]
line-length = 120

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "django5course.settings"
